# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CreateShareRequest:
    properties:
      encryptedCollectionKey:
        type: string
      encryptedCollectionKeyNonce:
        type: string
      expiresInHours:
        type: integer
      shareType:
        type: string
      targetId:
        type: string
    type: object
  handlers.CreateShareResult:
    properties:
      id:
        type: string
      token:
        type: string
    type: object
  handlers.DownloadURLResponse:
    properties:
      url:
        type: string
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.FileRow:
    properties:
      collectionId:
        type: string
      createdAt:
        type: string
      encryptedFileKey:
        type: string
      encryptedMetadata:
        type: string
      encryptedSizeBytes:
        type: integer
      fileKeyNonce:
        type: string
      id:
        type: string
      metadataNonce:
        type: string
      updatedAt:
        type: string
      uploaderUserId:
        type: string
    type: object
...
  handlers.PublicShareResponse:
    properties:
      encryptedCollectionKey:
        type: string
      encryptedCollectionKeyNonce:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      shareType:
        type: string
      targetId:
        type: string
    type: object
...
host: localhost
info:
  contact: {}
  description: Self-hosted, end-to-end encrypted file storage with federation. All
    file content and metadata are encrypted client-side; the server stores only ciphertext.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Kutup API
  version: 1.0.0
paths:
...
  /share:
    post:
      consumes:
      - application/json
      parameters:
      - description: Share details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateShareRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateShareResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a public share link
      tags:
      - Public Shares
  /share/{token}:
    get:
      parameters:
      - description: Share token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PublicShareResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "410":
          description: Link expired
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get public share metadata
      tags:
      - Public Shares
  /share/{token}/download/{fileId}:
    get:
      parameters:
      - description: Share token
        in: path
        name: token
        required: true
        type: string
      - description: File UUID
        in: path
        name: fileId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DownloadURLResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "410":
          description: Link expired
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Get presigned download URL for a public share file
      tags:
      - Public Shares
  /share/{token}/files:
    get:
      parameters:
      - description: Share token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.FileRow'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "410":
          description: Link expired
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: List files in a public share
      tags:
      - Public Shares
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"